Put vm specific logic into separate files
