+++ express_relay/sdk/js/src/svm.ts
@@ -1,0 +1,135 @@
+import {
+  Connection,
+  Keypair,
+  PublicKey,
+  Transaction,
+  TransactionInstruction,
+} from "@solana/web3.js";
+import * as anchor from "@coral-xyz/anchor";
+import { BidSvm, ExpressRelaySvmConfig } from "./types";
+import { AnchorProvider, Program } from "@coral-xyz/anchor";
+import { ExpressRelay } from "./expressRelayTypes";
+import expressRelayIdl from "./idl/idlExpressRelay.json";
+import { SVM_CONSTANTS } from "./const";
+import NodeWallet from "@coral-xyz/anchor/dist/cjs/nodewallet";
+
+function getExpressRelayProgram(chain: string): PublicKey {
+  if (!SVM_CONSTANTS[chain]) {
+    throw new Error(`Chain ${chain} not supported`);
+  }
+  return SVM_CONSTANTS[chain].expressRelayProgram;
+}
+
+export function getConfigRouterPda(
+  chain: string,
+  router: PublicKey
+): PublicKey {
+  const expressRelayProgram = getExpressRelayProgram(chain);
+
+  return PublicKey.findProgramAddressSync(
+    [Buffer.from("config_router"), router.toBuffer()],
+    expressRelayProgram
+  )[0];
+}
+
+export function getExpressRelayMetadataPda(chain: string): PublicKey {
+  const expressRelayProgram = getExpressRelayProgram(chain);
+
+  return PublicKey.findProgramAddressSync(
+    [Buffer.from("metadata")],
+    expressRelayProgram
+  )[0];
+}
+
+export async function constructSubmitBidInstruction(
+  searcher: PublicKey,
+  router: PublicKey,
+  permissionKey: PublicKey,
+  bidAmount: anchor.BN,
+  deadline: anchor.BN,
+  chainId: string,
+  relayerSigner: PublicKey,
+  feeReceiverRelayer: PublicKey
+): Promise<TransactionInstruction> {
+  const expressRelay = new Program<ExpressRelay>(
+    expressRelayIdl as ExpressRelay,
+    {} as AnchorProvider
+  );
+
+  const configRouter = getConfigRouterPda(chainId, router);
+  const expressRelayMetadata = getExpressRelayMetadataPda(chainId);
+  const svmConstants = SVM_CONSTANTS[chainId];
+
+  const ixSubmitBid = await expressRelay.methods
+    .submitBid({
+      deadline,
+      bidAmount,
+    })
+    .accountsStrict({
+      searcher,
+      relayerSigner,
+      permission: permissionKey,
+      router,
+      configRouter,
+      expressRelayMetadata,
+      feeReceiverRelayer,
+      systemProgram: anchor.web3.SystemProgram.programId,
+      sysvarInstructions: anchor.web3.SYSVAR_INSTRUCTIONS_PUBKEY,
+    })
+    .instruction();
+  ixSubmitBid.programId = svmConstants.expressRelayProgram;
+
+  return ixSubmitBid;
+}
+
+export async function constructSvmBid(
+  tx: Transaction,
+  searcher: PublicKey,
+  router: PublicKey,
+  permissionKey: PublicKey,
+  bidAmount: anchor.BN,
+  deadline: anchor.BN,
+  chainId: string,
+  relayerSigner: PublicKey,
+  feeReceiverRelayer: PublicKey
+): Promise<BidSvm> {
+  const ixSubmitBid = await constructSubmitBidInstruction(
+    searcher,
+    router,
+    permissionKey,
+    bidAmount,
+    deadline,
+    chainId,
+    relayerSigner,
+    feeReceiverRelayer
+  );
+
+  tx.instructions.unshift(ixSubmitBid);
+
+  return {
+    transaction: tx,
+    chainId: chainId,
+    env: "svm",
+  };
+}
+
+export async function getExpressRelaySvmConfig(
+  chainId: string,
+  connection: Connection
+): Promise<ExpressRelaySvmConfig> {
+  const provider = new AnchorProvider(
+    connection,
+    new NodeWallet(new Keypair())
+  );
+  const expressRelay = new Program<ExpressRelay>(
+    expressRelayIdl as ExpressRelay,
+    provider
+  );
+  const metadata = await expressRelay.account.expressRelayMetadata.fetch(
+    getExpressRelayMetadataPda(chainId)
+  );
+  return {
+    feeReceiverRelayer: metadata.feeReceiverRelayer,
+    relayerSigner: metadata.relayerSigner,
+  };
+}
