--- express_relay/sdk/js/src/svmPda.ts
@@ -1,30 +1,0 @@
-import { PublicKey } from "@solana/web3.js";
-import { SVM_CONSTANTS } from "./const";
-
-function getExpressRelayProgram(chain: string): PublicKey {
-  if (!SVM_CONSTANTS[chain]) {
-    throw new Error(`Chain ${chain} not supported`);
-  }
-  return SVM_CONSTANTS[chain].expressRelayProgram;
-}
-
-export function getConfigRouterPda(
-  chain: string,
-  router: PublicKey
-): PublicKey {
-  const expressRelayProgram = getExpressRelayProgram(chain);
-
-  return PublicKey.findProgramAddressSync(
-    [Buffer.from("config_router"), router.toBuffer()],
-    expressRelayProgram
-  )[0];
-}
-
-export function getExpressRelayMetadataPda(chain: string): PublicKey {
-  const expressRelayProgram = getExpressRelayProgram(chain);
-
-  return PublicKey.findProgramAddressSync(
-    [Buffer.from("metadata")],
-    expressRelayProgram
-  )[0];
-}
